rename main snaps settings page
