+++ app/components/Views/Snaps/SnapsSettingsList.tsx
